Feature-53 Add ViewModels for Tickers, Bought Assets, Recommendations, Buy, and Sell
